From: Pat Rankin <rankin@eql.caltech.edu> Subject: patch: engulfer explusion vs level change
     Fix a rare but severe problem that came up in the newsgroup
not too long ago.  If an engulfing monster expels the character
onto a level teleporter or trap door while the monster is moving,
then the resulting level change works ok because it doesn't happen
until monsters finish moving.  But if an engulfer does that during
the player's turn (zapped by a wand of opening, for instance), the
level change would occur immediately, the monster would get saved
and discarded with the old level, then a stale pointer to it would
be accessed while finishing the player's current move.  Assorted
very bad things could happen as a result.