From: michael.allison@mnr.gov.on.ca To: nh-patches@balder.sabir.com Subject: patch: invis/mummy wrapping bug
I think this bug has been in there since 3.2.2, but
I just discovered it during playtesting.

If you are wearing a mummy wrapping, and you zap yourself
with a wand of make invisible you incorrectly get the message
"Gee, all of a sudden you can't see yourself", which is
wrong since the effect never really happens.

Rather than just supressing the message, however,
this patch actually makes the mummy wrapping
absorb the ray and prevents it from affecting you.

In other words, the mummy wrapping protects you
from the effect (an invisibility dampening
field sort of effect) which is consistent with
what a mummy wrapping does to you too.

The rest of the changes are to replace
several instances of You("feel...
with You_feel("...