From: Pat Rankin <rankin@eql.caltech.edu> Subject: patch: several bits
     Several one-liners:

1) Add "rider/riders" as an index to the famine/pestilence/death
`whatis' entry, since that's how they're referred to in the code
and how various players refer to them in the newsgroup.

2) Replace "steel boots" with "buckled boots" since the description
won't necessarily end up being applied to metallic shoes.

3) Add "master lichen" as a new hallucinatory monster.  I don't
remember who submited lichen and master lichen as monsters, but
the latter is too cute to omit once the former has been adopted.