From: Mike Stephenson <mike.stephenson@spectrasoftware.com> Subject: FW: Patch: Guidebook.mn (first pass)
Here's my first kick at the fixes to Guidebook.mn.  I'll provide a patch to
Guidebook.tex when this one is complete.
I've got to do some research into the command line / .rc options and add
them in to complete the update, but I'd like to get this one out tonight in
order do deal with the game-related feedback.
I'll take a crack at adding the extmenu stuff in as well in the next pass.