From: Mike Stephenson <mike.stephenson@spectrasoftware.com> Subject: Patch: Update nethack.6
I've made an update to nethack.6 to cover some of the new command line
options.  While doing so, I've deliberately removed reference to the older
"-role"
single letter options for player role selection.  I've also noted (follows
in the next posting) a discrepancy between the "scoreboard" and "play" modes
of invocation surrounding the use of roles and races.  I doubt we'll be able
to fix it in time for the release, but it's a relatively low priority so
long as it's documented.