From: Ken Arromdee <arromdee@rahul.net> Subject: patch: beta-tester stuff
Fix some things noted by the beta-testers:
-- / on the player should not say "human or elf" if it includes orcs and
dwarves.  There is still no way to get this absolutely right as long as some
dwarves are 'h' and player dwarves are still '@'--what we really should do is
get the proper o or h symbol for the player and leave the description as
"human or elf".  But this is at least useful as a stopgap.
-- No rusted or otherwise damaged armor ever had the rust/damage information
printed.  Yeah, this one was my fault. :-)
-- 'q' should not be an option to the question asking if you want to pray.
-- eating a fortune cookie and reading the message should count as literacy,
since reading the message using the 'r' command counts.