From: Ken Arromdee <arromdee@rahul.net> Subject: patch: invisible monsters in corridors
Pat Rankin reported a problem where killing an invisible monster in a corridor
while blinded caused the square on the screen to be displayed as a blank
square instead of a corridor.

The problem can be reproduced as follows:
-- magic map the level (wizard mapping is okay)
-- put on a blindfold
-- teleport into a corridor
-- manually move a square down the corridor
-- create some monsters, and kill them
The behavior is that killing the invisible monster on the square where you
manually moved from does not display a blank, but killing the invisible
monster on the square where you had never moved does create a blank--in other
words, the real problem was with magic mapping, not invisibility, since magic
mapping creates squares that behave differently from squares that are actually
walked over.

I don't believe that this patch has any side effects, and none were obvious,
but anyone who thinks otherwise, please tell me.