From: Ken Arromdee <arromdee@rahul.net> Subject: patch: long worm beta-test fix, Guidebook
Fix the "aesthetics" mistake that crawled into the Guidebook from AD&D.  Yes,
I checked a dictionary to confirm it.

Fix the beta-test problem with getting an 'I' when attacking a long worm.  This
wasn't an invisibility bug, it was a long worm bug.  The bug is that the
function which decided if you can see a worm started at the second segment
of the worm.  If the only segment on-screen was the first segment (which is
the tip of the tail, since the list is backwards compared to the worm), this
would falsely report that you can't see the worm (thus triggering an
invisibility display.)

I don't know if this bug was present in 3.2.  If so, it needs to be added to
fixes33.0.  If the bug was present in 3.2, it would probably just give "You hit
it."