From: Pat Rankin <rankin@eql.caltech.edu> Subject: patch: Guidebook update
     Either I've remembered TeX's hyphen handling wrong, or
LaTeX handles it differently from plain TeX.  Anyway, almost all
the hyphens and dashes in the TeX edition of the Guidebook were
ok, although some sentence dashes were using the number range
dash instead.

     This cleans up various things that I noticed while looking
at the dashes.  It does not address anything pointed out by the
beta testers, unless by coincidence (such as adding monk and
ranger to the `name' option).  The TeX version retains one
Amiga-specific sentence (alt-meta stuff) as a comment (ie, still
in the source but won't appear in the formatted output).  I don't
know how to do comments in nroff so it got deleted from that
version.

     This time the TeX version has been tested and the nroff
one has not, the opposite of what's usually happened in the past.
LaTeX complains about the first sentence of the `dungeon' option
(which contains a long string of the default map characters), but
I did not attempt to fix that.  The printed version ends up with
"##" in the right margin and then breaks the line at what looks
like was intended to be a non-breakable space.  Someone should
verify that all the intended symbols do actually show up....

Files patched:
  doc/Guidebook.mn, Guidebook.tex