From: Pat Rankin <rankin@eql.caltech.edu> Subject: patch: goddess vs god
B1046:
     Implement the suggestion that the initial legacy message use
"goddess" instead of "god" when the deity is female.  I had intended
to do that a long time ago and just forgot about it.  The list of
deity names now uses a leading underscore to flag females.  At least
it flags those already described as female in data.base; a bunch of
the gods do not have an entry, and I made no attempt to dig up other
reference info on them.

     Also other minor quest.txt bits reported by Dylan:
Sam 00070 said "sun god" even though data.base describes Amaterasu
  Omikami as "sun goddess".
Kni 00069 had "it shalt" but "it" is 3rd person and the dictionary
  says that "shalt" is archaic 2nd person "shall".

Files patched:
  dat/quest.txt
  src/pray.c, questpgr.c, role.c