From: Ken Arromdee <arromdee@rahul.net> Subject: patch: beta-test
Fix the trivial colored glass bug bug (B1053).
Make wands of enlightenment self-identify when engraved (B1033).
Fix a problem that _might_ have caused the altar to display "There is
an  here." (B1052).  The altar name function printed into a non-static buffer
and tried to return a pointer to it.  I can't test this fix since it works
okay for me anyway, but this certainly _could_ have caused the problem.

Modified Files:
	src/invent.c src/objnam.c src/zap.c