From: Ken Arromdee <arromdee@rahul.net> Subject: patch: beta-test
Fixing the following items:
-- Missing period in "Your left hand is empty" (B2007)
-- Sink problems with levitation rings (B2004)
-- applying pick-axe while wielding two weapons doesn't immediately give a
warning message (B1072).  Note: I also added an is_weptool() check, and a
pick-axe is one-handed, so now it gives no message at all.  The bug also
existed for polearms, and the fix is actually applicable there.  If it turns
out the intention was not to allow weptools with two weapons, I'll change that
part back.
-- typing a return when asked what to call something, then a ^P, printed
garbage (B1067)--caused by terminating a buffer only after text was added to
it.  This seemed to be a tty-only bug.

Also reworded a message so it doesn't say you can't use two weapons when
"carrying" a shield (you have to wear it, not carry it).

Modified Files: src/apply.c src/hack.c src/wield.c win/tty/getline.c