From: Pat Rankin <rankin@eql.caltech.edu> Subject: patch: varlet
+B1069 "Methinks thou ..., eh, varlet?" malediction wraps due to length (Dylan)

     This was stranger than the impression I got from the original
description:  the line up through ``varlet'' was printed and triggered
``--More--'' on the second line of the screen.  Then after acknowledging
that, you'd get ``?"'' on the new top line, as if it had been a separate
message.

     The shorter variation still puts ``--More--'' on the second line
just like any other long message, but the full text fits on the first
line.  I didn't try to figure out why the original doesn't produce
``?"--More--'' on the second line.