From: Pat Rankin <rankin@eql.caltech.edu> Subject: patch: scoring fix
+B20?? adding artifacts to final score triggers a crash (Bruce)

     It doesn't happen in wizard mode because no score gets calculated
in that case.  While testing the fix, I escaped the dungeon with Orcrist
and Excalibur; it told that Orcrist was worth 40 zorkmids and Excalibur
was worth 60.  These values are ridiculously low, matching how much a
shopkeeper will offer to buy them for.  The corresponding sale price
is 25 times higher and would be more appropriate for scoring purposes.
I suspect that this feature is just going to point out how arbitrary
the shop prices are, because the actual capabilities, hence true value,
of artifacts gets completely ignored (by shks as well as by scoring).