From: Pat Rankin <rankin@eql.caltech.edu> Subject: patch: multishot
+B20?? provide better control over "multi-shot" missile launching (devteam)

     Rangers, samurai, elves, and characters with advanced weapon
skill in bow/crossbow/sling now shoot multiple missiles with a
single "throw".  But the player has no control; if the first shot
kills the intended target, any subsequent ones will hit whatever
is behind it.

     This patch doesn't stop you from shooting when the target
is killed--that might not be desireable when multiple targets are
all lined up.  What it does do is change the way a command count
prefix is interpreted for the 't' and 'f' commands.  Instead of
being a relatively useless repeat count, it will be a limit on
the number of shots that the multi-shot capability will take.

     I still need to update the Guidebook to describe the new
behavior, but I want to get the code in place now.