From: Pat Rankin <rankin@eql.caltech.edu> Subject: patch: artifact scoring
     Shopkeepers multiply the artifact cost twice, 4x in one spot
which Ken evidently found when adding artifacts to the score, and
25x in another, which he evidently missed.  This makes the value
at end of game correspond:  100x instead of 4x.  The point amount
is not increased by proportionally as much, 250x instead of 40x.
The formula still makes Giantslayer be worth as much as Frost Brand
or Excalibur since they're all long swords, which is absurd, but I
don't think there's sufficient time to do much about that.