From: Pat Rankin <rankin@eql.caltech.edu> Subject: patch: tuning
     When making normal levels, the usual number of monsters are
created even if the types normally chosen have been genocided.
But on special levels, there has been no substitution to fill in;
if the monster it wants to create has been genocided, there won't
be any monster.  This patch lets it replace the lost monster with
a random one.

Modified Files:
	doc/fixes33.0 src/sp_lev.c