From: michael.allison@mnr.gov.on.ca Subject: patch: config file change and documentation
Change the name of the config file for DOS and NT from
nethack.cnf to defaults.nh (for DOS) and Defaults.Nethack-settings
(for NT).  Also bring documentation for NT and DOS up-to-date
(I could swear I did this already once.)
Background: The speed-dialer program for Windows 9x and NT uses the
file extension .cnf.  This means that on some machines the
NetHack.cnf was hidden at worst, or labeled as a "speed-dial"
program at best.

After this patch:

On the DOS port, if there is no defaults.nh file, but there is
a nethack.cnf, then it will be used (for backward compatibility).

There is no rename of sys/share/nethack.cnf required in the
distribution, the change is accomplished through the Makefiles
when copied to the playground.

Modified Files:
	doc/Guidebook.mn doc/Guidebook.tex doc/nethack.6 src/files.c
	sys/msdos/Install.dos sys/msdos/Makefile.BC
	sys/msdos/Makefile.GCC sys/msdos/Makefile.MSC
	sys/msdos/Makefile.SC sys/msdos/msdoshlp.txt
	sys/msdos/pcvideo.h sys/msdos/template.mak sys/msdos/video.c
	sys/share/NetHack.cnf sys/winnt/Install.nt
	sys/winnt/Makefile.nt sys/winnt/winnt.cnf