From: Pat Rankin <rankin@eql.caltech.edu> Subject: patch: remove demihuman
     Change the description of '@' from "human/demihuman" back to
its original "human or elf".  ('/' will still append "or you" when
you look at yourself.)

     Also, a couple of DECgraphics bits:  iron bars had the value
``252 /* small pi */'' but that is actually the non-equals sign;
change it to 251 so that it really is the pi symbol.  (The diagonal
cross bar of not-equals is so thin that's is hard to see, so the
not-equals symbol isn't a good one to use.)  While in there, make
ladders use `>=' and `<=' instead of duplicating the stair symbols.

Modified Files:
	src/drawing.c