From: Ken Arromdee <arromdee@rahul.net> Subject: patch: beta-test
Fix B2049 (monster can flee with you inside it).  The solution is not to let
it flee, since that's a lot quicker a solution for me to figure out.  (The rest
is just indentation.)

Fix B3003 (minor documentation reference to 3.2).

Fix bug reported by Bruce that's not in the beta list yet: buzz() does a m_at()
before checking to see if the location is okay.

Modified Files:
	doc/window.doc src/uhitm.c src/zap.c