From: Ken Arromdee <arromdee@rahul.net> Subject: patch: beta-test
This fixes B3022 (blessed monster detection becomes identified even if there
are no monsters) and B3014 (potion thrown by monster angers monster).

The invisibility/I bug (B3015) is really annoying me.  I can't figure out
anything wrong with the code, nor can I reproduce it (except for the cases
that I already fixed a while back--long worms and magic mapping).

The code in mhitu.c puts up the 'I' if !canspotmon(mon).  The code in do_name.c
prints "it" if !canspotmon, if a few special cases aren't true, and finally if
    if (!mon_visible(mtmp) || (!cansee(bhitpos.x,bhitpos.y) &&
    !see_with_infrared(mtmp)))
mon_visible refers to invisibility (etc.) and is normally true even if you
are blind.  see_with_infrared() is always false if you are blind (it's also
redundant since !canspotmon implies !see_with_infrared, but I'm not changing
that right now), so the remaining suspicious part of this is the bhitpos part.
While I'm getting a headache trying to figure out exactly when this comes into
effect, this appears to do the right thing even if you're not hitting
anything: hack.c sets it every turn to your own position, and when blind
you're not capable of seeing your own position.

Modified files:
	include/extern.h src/dothrow.c src/mthrowu.c src/potion.c
	src/uhitm.c