From: arromdee@rahul.net Subject: patch: gas spore solved
I managed to reproduce the gas spore problem.  After further testing and
thinking, I figured out a problem.  Before gas spores were introduced, there
were no monsters that could cause damage at exactly that point in the turn.

Gas spores could damage pets just before grow_up() is called.  Since growing
up brings with it an increase in hit points, it is possible for that increase
in hit points to drive the pet from 0 hp to 1 hp, thus bringing it back to
life.  (The way the calculation works there is only a 1/4 chance of this
happening even if the dog is at exactly 0 hp.)

Modified Files:
	src/makemon.c