From: michael.allison@mnr.gov.on.ca Subject: patch: MSDOS overlay fix
This fixes the real mode overlay DOS level-file bug.
It wasn't a compiler bug, or an overlay thing, and it was my fault.
The reason it doesn't show up in the djgpp build is because it
doesn't define MFLOPPY. [ Looking for levels.* files, when the
level files are named playername.* doesn't work too well]

Also removes the gcc switches from the debugging flags of MSC. How
they got in there I have no idea.  They may have been in there in
3.2.2, I didn't check, and I don't do the debugging on the
overlaid version, except when its a bug that is only showing up there.

Modified Files:
	doc/fixes33.1 src/files.c sys/msdos/Makefile.MSC