From: arromdee@rahul.net Subject: patch: tins
I was perfectly able to both reproduce the tin problem and find the problem.
You need to answer "y" to the question as to whether to eat the tin off the
ground but "n" after opening it.

The lichen problem is still there.  The cause is obvious--eating tins
unconditionally sets u.uconduct.flesh, but the problem with this is that I
have no idea what was intended.

Someone with the flesh flag but not the meat flag is described as a vegetarian
but not a vegan by the code.  Unfortunately, the meaning of these words (a
strict vegan does not eat milk or eggs) does not fit the comments.  So I
have no way to know which part to fix.  Kevin?
	long	meat;		/* eaten corpse, tin, egg, tripe, body part */
	long	flesh;		/* ... or anything "fleshy" */

Modified Files:
	doc/fixes33.1 src/eat.c