From: Pat Rankin <rankin@eql.caltech.edu> Subject: patch: mummy wrappings worn by monsters
     A mummy wrapping was supposed to make an invisible monster
be visible just like for the player, but that hasn't worked right
since this feature was put in and no one reported it (until now
in the newsgroup).

     Testing this reveals a new buglet:  if you steal the worn
mummy wrapping and make a monster revert to invisible again, it
won't be displayed as 'I' for subsequent attacks in the same
turn.  (Ie, being polymorphed into an incubus or succubus who
gets multiple attacks and takes the cloak on the first one will
then hit "it" without any marker being put on the map.

Modified Files:
	doc/fixes33.1 src/worn.c