From: Dave Cohrs <cohrs@Legato.COM> Subject: patch: crashes from saving while mounted and other purify problems
Based on a bug reported today which stated that restoring a game that was
saved while mounted causes a sigsegv in weight_cap().

After some research, I believe this SEGV was a result of saving a bogus
m_id for the steed, which was caused by the fact that the steed monst* gets
freed before savegamestate is even called.  In fact, I got a failure during
save, not during restore.

Some further research showed that stuck monsters have the same problem; the
stuck problem existed back in 3.2.2 as well.

I "solved" this problem by saving the monster id's for stuck and steed
before calling savelev(), which eventually causes the monster data
structures to be freed, and modifying savegamestate to use these saved
values instead of the u.usteed or u.ustuck pointers.  The downside is that
there are two more static unsigned's cluttering things up just to make save
work.  I couldn't see an obvious way of avoiding this that wouldn't result
in memory leaks in some cases, but if someone wants to re-fix this, be my
guest.  After fixing this, I got no purify complaints or runtime errors
while saving or restoring while mounted.

While using purify to track this down, I found a freed-memory-read caused
by taming a monster with food, and a memory leak in the X11 port as
well.

Modified Files:
	doc/fixes33.1 src/dothrow.c src/save.c win/X11/dialogs.c
	win/X11/winX.c