From: arromdee@rahul.net Subject: patch: steeds, web page comments
SC7 on the web page was described wrong.  The actual bug is that if you have
Conflict and an aggressor monster dies in the Conflict, another monster could
still try to attack the first monster during the same turn even though it is
dead.

SC6 is also wrong.  Giants are *supposed* to have trouble picking up boulders
on sokoban.  The problem was that the message falsely said they didn't.

Bug report: hallucinating when dying from riding would print the
hallucination name instead of the real name of the monster on the tombstone.

Related but not actually reported to us: the message about hitting the ground,
that is printed for descending to the ground after levitating, is supposed
to be suppressed when dismounting.  However, under hallucination it wasn't
being suppressed.

Modified Files:
	doc/fixes33.1 src/steed.c src/trap.c