From: Pat Rankin <rankin@eql.caltech.edu> Subject: patch: twoweapon fix
     This fixes the reported bug that being granted an artifact
weapon when offering on an altar while in two-weapon combat mode
would unrestrict the two-weapon skill instead of the skill for
the artifact's weapon type.  The problem was more general:  if
you threw things while wielding two-weapons you trained your
two-weapon skill instead of the skill for whatever item was being
thrown.  (Throwing should probably be disallowed when wielding
two weapons, but this patch doesn't attempt to implement that.)

Modified Files:
	include/extern.h src/apply.c src/uhitm.c src/weapon.c