From: arromdee@rahul.net Subject: patch: invisibility
Fix some problems with 'I' and a message problem with wands of probing and
invisible monsters.  While we're at it, make a separate function for printing
a monster name that ignores 'it', so we no longer need the killer="" kludge
for monsters that aggravate players.  (Well, it's still kludgey, but not as
badly, and the kludginess is isolated to a small section of the code.)

Modified Files:
	doc/fixes33.1 include/extern.h src/do_name.c src/invent.c
	src/muse.c src/pline.c src/sounds.c src/zap.c