From: arromdee@rahul.net Subject: patch: rust
Black pudding problems.  It turned out the fix needed was rather bigger than I
had expected.

The core of the problem was that there was never a corrosion damage type
implemented.  Instead, black puddings were given AD_RUST and they were
being special-cased to be treated as corrosion instead of rust.  The special-
casing assumed that anything corrodeable was made of bronze, and it was not
done at all when attacking monsters and getting your weapon rusted/corroded
by hitting them.

The fix, of course, is to add a separate AD_CORRODE.  As a bonus, the armor-
rusting functions no longer need the permonst parameter since that was only
used for special-casing.

Modified Files:
	doc/fixes33.1 include/extern.h include/monattk.h src/mhitm.c
	src/mhitu.c src/monst.c src/uhitm.c