From: michael.allison@mnr.gov.on.ca Subject: patch: more gelatinous cubic bug nh330
Dean Luick [dean@ni.com] wrote:
>On Tue, 28 Dec 1999 michael.allison@mnr.gov.on.ca wrote:
>> Fix problem reported by Christian Bressler [gaston@cs.tu-berlin.de]
>> where a buffer could be used without being initialized.
>
>Assuming that 'ecount' is "a seen eaten count" then perhaps you should
>move ecount's increment to inside the "if (cansee(...))".

Actually ecount was meant to be an "engulfed" count by the look of it.
Unlike the eating part earlier in the routine, this routine provided no
feedback when you couldn't see the monster. This reworks it slightly
to allow audible feedback ("slurping sound" as in the top half), and
defers the cansee() check to the actual printing of the message
after the loop.

Also changes 'pline(buf)' to 'pline("%s", buf)' since it occurred to
me that the reported crash must have been related to having stray
'%' characters in the unitialized buffer.

Modified Files:
	src/mon.c