From: Janet Walz <jwalz@radix.net> Subject: X11 player selection (from Pat)
     This implements the missing bits of role selection for X11:
constrain race values to those acceptable for chosen class and
let user choose gender and alignment instead of always assigning
random choices.  When some of the criteria are specified on the
command line, via NETHACKOPTIONS, or from config file, then the
choices presented to the user for the other criteria are filtered
by the parts that have been specified.  It works fairly well for
the testing I've done, but I don't ordinarily use this interface
and no doubt haven't done enough testing overall.

     The code uses Janet's ok_role(), pick_role(), etc routines
which were sent out for testing shortly before the 3.3.0 release
and haven't been adopted into the code yet.  So to test this
you'll need to apply that patch too.  **Please do test it if you
use the X11 interface.**

     The key accelerators seem to work, although they allow
the user to pick things which aren't shown on the menu.  That
was expected as I wrote it, and I doubt that it really needs to
fixed.  If you set `male' in NETHACKOPTIONS, valkyrie won't be
listed among the roles.  But you can type 'v' and pick valkyrie
anyway (there's currently no way to do the same thing with the
mouse).  If you do pick that unlisted role, then when gender
selection gets done the male setting from the environment will
be overridden and you'll end up with a normal female valkyrie
character.  Similarly, if you specified 'lawful' then "elven"
won't be listed as a choice for race, but if you pick it via
the 'e' key, you'd end up being a chaotic elf (provided that
the previously chosen role allows elves; if not, choosing 'e'
would just pop another "choose the race" menu until you pick
something compatible with the current role).

     For roles, the first letter of the role name is the key to
use.  Either upper or lower case works, except for rogue which
requires 'r' and ranger which needs 'R', compatible with the
present tty menu for role selection.  That gets determined
dynamically from the roles[] array.  If future roles conflict
in the first letter, whichever one comes first in the array
will have lowercase as its key and whichever comes second will
have uppercase.  If a third role uses the same starting letter,
there will not be any accelerator key to choose it--you'll be
stuck with the mouse.  (If that ever happens, then arbitrary
single letter designations can be added as a new roles[] field.)

     Races, genders, and alignments are handled similarly,
although none of those categories has any first letter conflicts
so upper and lower case works for all their values.  For all
four types of selection, both '*' and '@' work as accelerator
for "random"; <ESC>, 'q', and 'Q' work for "quit".  That's
checked after the roles[], races[], etc letters, so if a role
or race (or gender or alignment :-) beginning with "Q" ever
gets added, the 'q' key would choose it instead of quitting.
<return> and <space> do not work for quit, because in this
case is means "quit the game" rather than "cancel whatever
command caused this menu to be brought up" (the way that those
keys work for the extended commands menu).

     [There must be some way to merge the four foo_key()
routines and use one of their currently ignored arguments to
distinguish which type of keystroke is wanted.]

Modified Files:
	include/winX.h win/X11/winX.c win/X11/winmisc.c