From: Pat Rankin <rankin@eql.caltech.edu> Subject: patch: region fix
     I don't know whether this fixes the reported problem of the same
monster being killed multiple times by a stinking gas cloud; I haven't
been able to reproduce that.  But there is an array bounds problem
when a region contains a multiple of 5 monsters and the last one is
killed by the region callback routine.  (That would access a `short'
beyond the end of malloc'd memory, which isn't likely to trigger the
reported symptom.)

Modified Files:
	src/region.c