From: arromdee@rahul.net Subject: patch: x_monnam
This was a quick fix to a problem where someone reported "an" Asmodeus reading
a scroll of teleportation.  It ended up being a fix of something that really
did need fixing for a while.  The logic in the monster naming function was
a little convoluted, and I tried to straighten it out.

I probably should eventually change the functions that are one-line x_monnam()
calls--why not use #defines?  The original reason for doing it this way was a
little weird: I was compiling this long ago on a slow 286, was working on the
monster-naming functions, and didn't feel like recompiling the entire source
tree since it would take forever on the machines of that time.  If I kept a
function with the same prototype and had it call my new function, I could avoid
this.

I tried a bunch of cases, but it's always possible I missed something; if
anyone has the time, please try it a bit.  (The Asmodeus problem I started this
to fix is fixed.)

Modified Files:
	doc/fixes33.1 include/decl.h include/extern.h src/do_name.c
	src/hack.c src/mon.c src/muse.c src/music.c src/shk.c