From: Dave Cohrs <cohrs@mail.legato.com> Subject: patch: blindness
At long last, here's a patch (which I completely reworked after Michael
sent out that detailed list of ways to avoid certain kinds of blindless)
for the problem originally reported as Sunsword protects against raven
attacks.

A new function, can_blnd() is introduced.  This function bases its checks
on static and dynamic information, so it takes a bunch of parameters, but
the upside is it allows all the blindless logic to move to one place.  I
didn't replace all of the resists_blnd() calls with can_blnd(), but this
should be possible.  can_blnd() makes a bunch of assumptions which should
be valid for current code (unless I messed up).  The nice thing is, if these
assumptions become invalid, we only have to change one place.

I fixed a few places where blindness was handled inconsistently between
monster and you, and generally followed the outline in Michael's message (I
disagreed with the blanket use of "already blind", and applied it more
sparingly than Michael suggested; you may also notice that engulfing/dust
affects one more than Michael suggested).  One place in uhitm.c required a
major re-arrangement, but I think the logic works the same, but using the
new call.

Finally, I changed the AD_BLND to an AD_PHYS for the ki-rin, and gave it a
new AT_MAGC attack, both to be more consistent with the MM (I'm not sure if
AT_MAGC was around when the ki-rin was originally added; I think if it
were, I would have added this attack in the first place).  This results in
the ki-rin being a bit nastier than before, which is probably a good thing.

I also fixed at least one possible overflow condition.

Modified Files:
	doc/fixes33.1 include/extern.h src/mhitm.c src/mhitu.c
	src/mondata.c src/monst.c src/mthrowu.c src/read.c src/uhitm.c