From: Pat Rankin <rankin@eql.caltech.edu> Subject: patch: quiver
     I used the (Q)uiver command as a tourist and the prompt offered
choices of camera, credit card, and magic marker.  It was using the
same object class list as (w)ield but not filtering the tools down to
weapon-tools.  Even weapon tools aren't likely candidates for throwing,
so use a weapon-only object class list.  (The patch is bloated a little
bit due to some whitespace cleanup.)

Modified Files:
	doc/fixes33.1 src/read.c src/wield.c