From: arromdee@rahul.net Subject: patch: medusa
This fixes a medusa problem noted on the newsgroup.  It also corrects some
odd tabbing in uhitm.c (no change to code otherwise) and fixes a bug which
allowed medusa to drink a potion of invisibility.

The message is changed from "you look upon" to "you meet the gaze of", because
most of the rest of the code assumed that it was a gaze attack.

Modified Files:
	doc/fixes33.1 src/mhitu.c src/muse.c src/uhitm.c