From: Pat Rankin <rankin@eql.caltech.edu> Subject: patch: shop fix
     From the newsgroup:  dipping an object into a potion of polymorph
while standing in a shop would cause the shopkeeper to go ballistic.
The poly_obj() code assumed that only objects on the floor or inside
a container (which I don't understand, unless it matters to stone-to-
flesh) could be polymorphed so assumed that it was shop inventory that
was being trasmuted.

Modified Files:
	doc/fixes33.1 src/zap.c