From: Dean Luick <dean@ni.com> Subject: patch: fix monster leaves trail in lit room (REPLACEMENT)
This patch fixes the problem where if the hero reads a scroll of magic
mapping while engulfed, then there are cases where lit room locations
will go dark as a monster walks over them.

The basic problem is that map_background() assumes that the hero
can see the location.  So previously unseen lit room and corridor
locations were displayed as lit.  This was OK when not engulfed,
because the newsym() call immediately following map_background fixes
up those locations so they go dark.  When engulfed, the newsym call
is short circuited, so the fixup didn't occur.  So when a monster
walked over one of these spot, newsym would finally fix up the
visual, and the spot would go dark.

This fix intruces a magic_map_background function which will
set the correct visual whether engulfed or not.

Modified Files:
	doc/fixes33.1 include/extern.h src/detect.c src/display.c