From: michael.allison@mnr.gov.on.ca Subject: patch:
Started out to just vary the #chat message for a vampire
a little, especially when one vampire expressed a desire
suck another vampire's blood.

Ended up:

- Creating a new body_part type BLOOD, to be consistant with other
  body_parts when it comes to things like polymorphing etc.
- That meant changing engraving's use of BLOOD to ENGR_BLOOD.
- Along the way, I notice that the body_part macros (HAND, FOOT, etc.)
  were defined in you.h, even though they are now also used by
  mbodypart() so I moved them to the more general location, hack.h.
- varied the vampire messages depending on tameness, what type of
  monster you are, time of night, and peacefulness.

BTW, I realize that "life force" is lame and incorrect for
a sphere and a vortex, but blood is wrong too. If anyone has
a better term please patch it.

Modified Files:
	include/engrave.h include/hack.h include/you.h src/engrave.c
	src/polyself.c src/sounds.c