From: Pat Rankin <rankin@eql.caltech.edu> Subject: patch: returning boomerang
     Make a returning boomerang behave more like Mjollnir:  re-equipped
into the weapon slot it was in at the time it was thrown.  This is just
a convenience for the player; the only effect on game play would be if
it was in the secondary weapon slot and the wielded weapon was cursed;
the old behavior would prevent anything from being put into the emptied
secondary weapon slot.

Modified Files:
	doc/fixes33.1 include/extern.h src/dothrow.c src/polyself.c