From: Dean Luick <dean@ni.com> Subject: patch: flying iron balls and iron bars
This fixes the reported bug of kicked iron balls pass through iron bars.
We do this by adding a check in bhit() if the given object has the ability
to pass through iron bars.

The actual change to bhit() is small, the rest of it is noise from
me re-indenting it.

Modified Files:
	doc/fixes33.1 src/zap.c