From: Pat Rankin <rankin@eql.caltech.edu> Subject: patch: two-weapon & martial arts bits
     Someone in the newsgroup pointed out that being restricted
in two-weapon combat was worse than being unskilled when that is
not the case for any other skill (except that it would been the
case for bare-handed combat too, but no one is restricted in that
skill).  This removes that inconsistency, and also boosts the
martial arts damage bonus a little bit.

Modified Files:
	src/weapon.c