From: Dean Luick <dean@ni.com> Subject: patch: infrared monsters and vision
This patch fixes the reported problem where a monster seen via infrared
is not removed from the screen when the hero moves around a corner.
This also fixes the unreported problem of not seeing an unmoving
monster via infrared when the hero moves around a corner.

For example, play as an elf character and create a red mold in a dark
corridor.  Now move around a corner.  The mold should disappear.  Now
move back around the corner.  The mold should reappear.

Most of this patch is me adding comments (vision.h), minor reformatting
(display.c) and both (vision.c).  Sure, & binds tighter than &&, but I
just like those parenthesis.

The actual change is in the "not_in_sight" case (the last band) in vision.c.
What I'm doing is forcing a newsym call if a position is not IN_SIGHT and
either the new or old COULD_SEE bit is set, but not both.  That is,
a newsym call is now done any position falls into or out of being
seen-if-only-it-were-lit.

Modified Files:
	doc/fixes33.1 include/vision.h src/display.c src/vision.c