From: Dave Cohrs <cohrs@Legato.COM> Subject: patch: more on that buffer overrun
>Somebody did almost the same thing to wintty.c recently:
>:    const char *newstr;
>:    char buf[4+BUFSZ];
>
>and didn't include a comment about the 4 there either. ;-)

This is exactly the code I looked at before simply adding that 32+.

Anyway, here's more on that same vein (not a replacement), that stops
mis-using BUFSZ with toplines, adds the comment Janet wanted in this case,
and partly addresses the sys/amiga problem (although it doesn't address why
sys/amiga/winext.h re-declares all sorts of stuff that's already in decl.h
and other common include files, which I think is just wrong).  I replicated
the strncpy code from the amiga port to the tty and X11 ports just in case.

Modified Files:
	include/global.h src/decl.c sys/amiga/winext.h
	sys/amiga/winstr.c win/X11/winX.c win/tty/topl.c