From: Dean Luick <dean@ni.com> Subject: patch: jump/hurtle fix
This patch prevents the hero from passing through iron bars via hurtle
and stone via jump (only when wearing Eyes of the Overworld).  In
particular:

For jumping:
	o allow monster locations to be chosen
	o allow wall/door spots to be chosen -- it the player wants to
		crash into it, let them
	o don't pass through iron bars

For hurtle:
	o don't pass through iron bars
	o allow passing through open doors

Both jump() and hurtle() now call a path walker that will check each
spot along the path for validity.  They use the same callback, and so now
have exact features (although this can be changed).

I put walk_path() in dothrow.c because it seems to be as good of a spot
as any other.

Modified Files:
	doc/fixes33.1 include/extern.h src/apply.c src/dokick.c
	src/dothrow.c