From: Pat Rankin <rankin@eql.caltech.edu> Subject: patch: polyself buglet--golem gender
     A quibble in the newsgroup complained that kicking sinks was
more likely to summon succubi than incubi, because neuter monsters
get lumped in with males.  The only neuter monsters that're able
to kick that I could see are the various flavors of golems, and it
turns out that since they're classified as humanoid, they were being
treated as male for seduction attacks.  This fixes that and also the
quibble that led to its discovery.

Modified Files:
	doc/fixes33.1 src/dokick.c src/polyself.c